+++ packages/riverpod_state_provider/riverpod_state_provider_annotation/lib/riverpod_state_provider_annotation.dart
@@ -1,3 +1,4 @@
+export 'src/helper.dart';
 export 'src/riverpod_state_provider_annotation.dart';
 
 // ignore: invalid_export_of_internal_element