+++ packages/riverpod_state_provider/riverpod_state_provider_generator/test/int_test.dart
@@ -7,32 +7,36 @@ void main() {
   test('Creates an equivalent of the StateProvider for an integer', () {
     final container = createContainer();
 
-    var value = container.read(myIntegerProvider);
+    var value = container.read(myIntegerStateProvider);
     expect(value, 0);
 
-    container.read(myIntegerProvider.notifier).state++;
-    value = container.read(myIntegerProvider);
+    container.read(myIntegerStateProvider.notifier).state++;
+    value = container.read(myIntegerStateProvider);
     expect(value, 1);
 
-    container.read(myIntegerProvider.notifier).update((state) => state + 1);
-    value = container.read(myIntegerProvider);
+    container
+        .read(myIntegerStateProvider.notifier)
+        .update((state) => state + 1);
+    value = container.read(myIntegerStateProvider);
     expect(value, 2);
   });
 
   test('The generated provider can be overridden', () {
     final container = createContainer(overrides: [
-      myIntegerProvider.overrideWithValue(42),
+      myIntegerStateProvider.overrideWithValue(42),
     ]);
 
-    var value = container.read(myIntegerProvider);
+    var value = container.read(myIntegerStateProvider);
     expect(value, 42);
 
-    container.read(myIntegerProvider.notifier).state++;
-    value = container.read(myIntegerProvider);
+    container.read(myIntegerStateProvider.notifier).state++;
+    value = container.read(myIntegerStateProvider);
     expect(value, 43);
 
-    container.read(myIntegerProvider.notifier).update((state) => state + 1);
-    value = container.read(myIntegerProvider);
+    container
+        .read(myIntegerStateProvider.notifier)
+        .update((state) => state + 1);
+    value = container.read(myIntegerStateProvider);
     expect(value, 44);
   });
 }