+++ packages/riverpod_state_provider/riverpod_state_provider_generator/test/integration/family.g.dart
@@ -7,7 +7,7 @@ part of 'family.dart';
 // **************************************************************************
 
 String _$myFamilyPositionalParametersStateHash() =>
-    r'eb3f9d771bef1e8a755bd46c1d72aefbd287bfe4';
+    r'3fbc7bfbef3ae193a1a9b34d9ed6152f80391e6c';
 
 /// Copied from Dart SDK
 class _SystemHash {
@@ -267,7 +267,7 @@ class _MyFamilyPositionalParametersStateProviderElement
 }
 
 String _$myFamilyMixedParametersStateHash() =>
-    r'51a2fd5d1ae0e1c1e5a3a371c508f8795a0fcd0c';
+    r'4e43255f17cb5fd1a05eb45ad05d48c1ccfb1820';
 
 abstract class _$MyFamilyMixedParametersState
     extends BuildlessAutoDisposeNotifier<List<bool>> {